+++ RELEASES.md
@@ -1,3 +1,10 @@
+# Version 0.1.2 (2021-03-12)
+
+  * Fix zero literal suffixes in [C11 implementation](c11).
+  * Use `num_f`, `num_d`, and `num_l` type definitions.
+  * Update build script.
+  * Update [README.md](README.md).
+
 # Version 0.1.1 (2021-03-11)
 
   * Reliably build documentation at <https://doc.qu1x.dev/trackball>.